rm hooks from graph utils
