+++ src/modals/ValidatorMetrics/index.tsx
@@ -8,8 +8,9 @@ import { useNetworkMetrics } from 'contexts/Network';
 import { useStaking } from 'contexts/Staking';
 import { useSubscan } from 'contexts/Subscan';
 import { EraPoints as EraPointsGraph } from 'library/Graphs/EraPoints';
-import { formatSize, useSize } from 'library/Graphs/Utils';
+import { formatSize } from 'library/Graphs/Utils';
 import { GraphWrapper } from 'library/Graphs/Wrappers';
+import { useSize } from 'library/Hooks/useSize';
 import Identicon from 'library/Identicon';
 import { Title } from 'library/Modal/Title';
 import { StatsWrapper, StatWrapper } from 'library/Modal/Wrappers';
