+++ src/pages/Overview/Payouts.tsx
@@ -6,7 +6,8 @@ import { useStaking } from 'contexts/Staking';
 import { useUi } from 'contexts/UI';
 import { PayoutBar } from 'library/Graphs/PayoutBar';
 import { PayoutLine } from 'library/Graphs/PayoutLine';
-import { formatSize, useSize } from 'library/Graphs/Utils';
+import { formatSize } from 'library/Graphs/Utils';
+import { useSize } from 'library/Hooks/useSize';
 import { StatusLabel } from 'library/StatusLabel';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
